Add Harvesting Memories short story
This is a short story idea I cam up with today durin a warmup for a writing get together. I had various ideas swirling around in my head, but the genre we chose was science fiction, and the idea to focus on was harvesting memories. The intro scene sees us meet our unnamed protagonist(?), and Imperium-trained memory hunter who is being sent on a mission to retrieve an unspecified object and some so-called 'memory banks' at the behest of their handler, Walker, on behalf of their ruler, Charon. The second scene bean us, in media res, into the moments in which the aforementioned protagonist and unspecified others are retrieving the object specified in the first scene. I was in more of a contemplative mood, so I wanted to start off by writing about some of the ideas that were on my mind, then gradually blending them into the scene in the form of a dialogue of sorts. An interesting twist here is that none of the characters are actually speaking, but are instead thinking, and I also don't know if they are humans yet. They exist in a world without written or spoken language; instead, their thoughts are connect, as if they were networked computers. In the scene, they are uncovering traces of a time far pre-dating them, one in which language was still spoken out loud and written down. They come across a shattered mirror and do not know what to make of it, hazarding their best guesses.